+++ examples/openai/smart_scraper_multi_openai.py
@@ -2,7 +2,8 @@
 Basic example of scraping pipeline using SmartScraper
 """
 
-import os, json
+import os
+import json
 from dotenv import load_dotenv
 from scrapegraphai.graphs import SmartScraperMultiGraph