+++ scrapegraphai/nodes/get_probable_tags_node.py
@@ -3,10 +3,8 @@ GetProbableTagsNode Module
 """
 
 from typing import List, Optional
-
 from langchain.output_parsers import CommaSeparatedListOutputParser
 from langchain.prompts import PromptTemplate
-
 from ..utils.logging import get_logger
 from .base_node import BaseNode